Massive cleanup + reorganization
