+++ source/c/gridrec.c
@@ -83,6 +83,9 @@ gridrec(const float* data, int dy, int dt, int dx, const float* center,
         const float* theta, float* recon, int ngridx, int ngridy, const char* fname,
         const float* filter_par)
 {
+    if(dy == 0 || dt == 0 || dx == 0)
+        return;
+
 #if defined(TOMOPY_CXX_GRIDREC)
     cxx_gridrec(data, dy, dt, dx, center, theta, recon, ngridx, ngridy, fname,
                 filter_par);
