+++ source/include/utils.h
@@ -62,7 +62,7 @@
 #endif
 
 #include "common.h"
-#include "gpu.h"
+#include "macros.h"
 
 //======================================================================================//
